REDUCE manual formatting
In applysym.tex, wrap a \small that was in a global context in {}; it
was affecting all of the manual after it.  Also, replace \[ \]
containing a \label with an equation environment so that the label is
not ignored.  Remove an unused equation label in crack.tex.  Improve
spacing in some equations in susy2.tex.

Other changes only affect the HTML version.

Delete obsolete file tex4ht.env, which was causing warnings (at least)
about missing font information.  Comment out the code in reduce-sf.cfg
that was intended to avoid the long (monospaced) space after \texttt{}
because it may lose the space completely, which is worse.  This needs
further investigation.  Revise internal hyperlinking within crack.tex,
gnuplot.tex, logoturtle.tex and susy2.tex; this worked (I think) in
PDF but not in HTML.  Fix use of math-mode commands in susy2.tex.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7127 2bfe0521-f11c-4a00-b80e-6202646ff360