ODESolve bug fix
Procedure newroot_of_unity_tag in module odetop calls mkrootsoftag
defined in module solve/solve1, but this might not be loaded.  So
instead include an updated version of the body of mkrootsoftag as the
body of newroot_of_unity_tag.  (It's very little code, so I don't
think the duplication is a problem.)  Revise procedure
odesolve!-bernoulli1 in module odenon1 to make the code slightly
clearer, but with no functional difference.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7130 2bfe0521-f11c-4a00-b80e-6202646ff360