REDUCE manual: HTML formatting update
crack.tex: \verb hack to avoid identifiers beginning with !* being
split over two lines between ! and * in the list of global variables.

susy2.tex: change the table of commands from 3 to 2 columns to avoid
the need to decrease the font size, which was not getting communicated
to MathJax.

reduce-sf.cfg: re-instate the code to avoid a long space after
\texttt, which seems to work again

symbolic.tex: replace \texttt{} around a table, which after the above
change caused invalid HTML, with \ttfamily.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7141 2bfe0521-f11c-4a00-b80e-6202646ff360