REDUCE manual: HTML version
Fix some inconsistent formatting and some that did not work in the
HTML version in the first paragraph of section 5.7.2 "Labels and GO TO
Statements".

Add an experimental script makehtml-for-sf.sh to make the HTML version
of the manual for use ONLY on the web site.  This script is equivalent
to 'mkhtml.sh for-sf' but it runs make4ht once instead of mk4ht twice.
It requires the (Lua) build script manual.mk4, which calls the
(Bash) script makehtml-index.sh to build the index.  It also requires
the non-standard software (in addition to TeX4ht) make4ht (in TeX Live
collection-binextra) and luaxml (in TeX Live collection-luatex).
makehtml-for-sf currently runs LaTeX 5 times and so is a little faster
than mkhtml, which runs LaTeX 6 times.  make4ht appears to have
replaced mk4ht as the preferred way to run TeX4ht.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7142 2bfe0521-f11c-4a00-b80e-6202646ff360