Web REDUCE
Commit some experimental JavaScript library code, and an updated
Makefile that targets the latest Emscripten (4.0.13) and
Node:(24.7.0), and REDUCE 7169.  It will not work with significantly
older versions.  It builds bootstrapreduce.{js,wasm} but it hangs when
run.  I have probably updated the CSL files wrongly!

I was able to build REDUCE 6762 that ran in Node (only) using a
similar Makefile (but with appropriate CSL files), although with any
optimization bootstrapreduce failed to run, apparently due to
alignment issues in arith01.cpp.  It would be best to explore this
issue using the latest version of CSL, once that can be built!


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7170 2bfe0521-f11c-4a00-b80e-6202646ff360