Web REDUCE Makefile
This version appears to fix the main problem with building Web REDUCE.
The key was to add the emscripten option -sMEMORY64, and also the CSL
option -D@package=... when running buildreduce.lsp.  I can now build a
fully optimized bootstrapreduce and node bootstrapreduce seems to
work, but I cannot build reduce.img because of errors compiling the
generated C files.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7171 2bfe0521-f11c-4a00-b80e-6202646ff360