REDUCE on Common Lisp
Put all fasl files in a directory specific to each version of Common
Lisp, which makes it easier to support more versions of Common Lisp
(such as Embedded Common Lisp):.  Make it clearer in the README that
Common Lisp must be obtained and installed independently (which avoids
unnecessary complications and licensing issues), and that building
within the Subversion file tree is very easy (for the versions of
Common Lisp that I fully support).  This revision builds SBCL REDUCE
on Windows, but I want to check on Linux and then try some better
testing.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7172 2bfe0521-f11c-4a00-b80e-6202646ff360