Common Lisp: build CCL REDUCE on macOS again, etc.
Strip experimental code out of build.sh.  Fix some customization
errors in sl-on-cl.lisp (thanks to Marco Ferraris).  Allow the redsbcl
commands to accept arguments that are passed on to SBCL.  Ensure
tracing output starts at the beginning of a line.  Try another
definition of equal, which works for sparse but not for ibalp, so it
is unlikely to be the final version.  Fix printing and exploding of
floats.  Fix gensym.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7177 2bfe0521-f11c-4a00-b80e-6202646ff360