Common Lisp: fix building on CLISP
Handle the ".lib" files generated by the CLISP compiler.  Revise
printing of floats to try to avoid unnecessary float computations,
which can cause either overflow or underflow.  Not tested very
thoroughly, but I can now build REDUCE on both CLISP and SBCL on
Windows, and my test results on SBCL are not changed.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7178 2bfe0521-f11c-4a00-b80e-6202646ff360