Common Lisp: equal and hash tables
Implement Standard Lisp EQUAL as Common Lisp EQUALP.  This is not
strictly correct but seems to be the best compromise.  Add support for
hash tables.  This leads to a speed increase on the standard test
files by a factor of about 3.  But there are still a few, mostly
insignificant, differences in the test output that I am investigating.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7183 2bfe0521-f11c-4a00-b80e-6202646ff360