Make gf2 work better on Common Lisp
rlisp/inter.red: allow with!-timeout to use resource!-limit if it is
defined, rather than only on CSL.  (I define this in SL-on-CL and it
sort-of works, at least on SBCL, where I call the SBCL with-timeout,
which has different semantics from the REDUCE version!)

gf2.red: make on backtrace the default, not only for PSL (although
this doesn't yet work on Common Lisp).

gf2groeb.red: define optterpri if it is undefined, rather than only
for PSL.

gf2.tst: set the print base to 16 for Common Lisp.

gf2.rlg: update (using CSL).


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7189 2bfe0521-f11c-4a00-b80e-6202646ff360