Revise some fluid/global statements
FLUID and GLOBAL (and related) declarations are supposed to be
evaluated even when the switch DEFN is ON (because they are flagged
EVAL).  But if they are preceded by LISP or SYMBOLIC then they are not
evaluated.  This causes a lot of spurious warnings about use of
undeclared variables when building REDUCE on Common Lisp (because I
use ON DEFN to generate Lisp files when compiling REDUCE files).
Therefore, I have deleted LISP or SYMBOLIC preceding FLUID and GLOBAL
(and related) declarations.  I believe this is safe because if the
first argument of a function is quoted then REDUCE automatically
evaluates it in symbolic mode (and the only case where an argument
was not quoted was already in symbolic mode).  This issue mostly
affects files in the "crack" directory, although it was "sstools" that
brought my attention to it.

Now I can see where there are real issues with the use of undeclared
variables.  The "plot" package produces such warnings because it uses
them as arguments to a macro, and "crack" still produces a couple
because it assumes that it is running on either CSL or PSL, which I
will fix later.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7199 2bfe0521-f11c-4a00-b80e-6202646ff360