Minor updates for Common Lisp
build.sh: Move the time stamp line earlier so that it works.  Better
handling of build errors: disable the interactive debugger when
building on SBCL; combine related bash commands using && so that
subsequent commands are not run if an earlier one fails.

sl-on-cl.lisp: Revise support for FLUID and GLOBAL.  Finally fix
handling of *gc-hook* on SBCL.  Flag bothtimes and compiletime eval,
which solves a problem shown up by a spurious warning when running
"sstools.tst".  Enable the interactive debugger before saving a REDUCE
image on SBCL.  Add the symbol sl-on-cl to lispsystem* (although I
don't currently use it).

REDUCE on SBCL now runs the standard set of test files with mostly
insignificant differences in output only for the arith, ellipfn, gf2
and numeric packages, some of which may be unavoidable.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7202 2bfe0521-f11c-4a00-b80e-6202646ff360