Common Lisp improvements
Revise getd to return a function pointer for a compiled function
instead of always returning a lambda expression.  Generalise explode
to handle function pointers.  Revise yesp to output its message with
better spacing and lower-case identifiers.  Enable the SBCL
interactive debugger only if the input and output are both
interactive.  Enable initial compilation on SBCL only if *comp is
true.  Remove eval-when from the definition of eqcar.  Revise some
comments.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7205 2bfe0521-f11c-4a00-b80e-6202646ff360