Common Lisp: CLISP REDUCE builds again
Revise getd to regard special operators and macros that are not
declared as SL macros as fexprs (more to the point, not as macros or
exprs).  This seems to be the key to building REDUCE on CLISP again.
(putd doesn't support fexprs, but this doesn't seem to matter!)  getd
and putd probably still need some more work.

Better debugging support.  Remove the need for quotes in arguments to
defalias.  Avoid all undefined function warnings when compiling
sl-on-cl.  Update README, especially the current status of REDUCE on
Common Lisp.  Set the character code for CLISP during the initial
build phase to be UTF-8.  Add redclisp.bat; update comments in the
redclisp, redccl and redccl.bat scripts.


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7207 2bfe0521-f11c-4a00-b80e-6202646ff360