The new directory csl/new-embedded/using-clang contains a Makefile that can go some way towards building a system on windows-on-arm if used under the mingw64-clang setup. There I have managed to make it build an executable that is in effect a bootstrapreduce.exe, and that starts to run a Reduce build bur crashes out very very early - and it is not clear that I have good debugging tools there that are in a state where I understand them. If I build on an x86_64 windows I can make the .exe file but if I try to run it it needs a bunch of libraries (presumably .dll files) that at present it does not know where to scan for. If I try linking statically I get an utter mess as libraries intended for g++ and clang++ seem to conflict.
This is one of the key motivations for the recent checkin regarding
"alignof". I have also tidied up some code so I do not get frivolous
moans about i32_to_f128 that bit me here and have hurt others too.
Plus other adjustments to cope with clang++ foibles visible here on
windows-on-arm but not manifesting on MacOSX.

It feels good to have got anywhere at all with Windows-on-ARM but this
is just a first step. However it may amuse others since using-clang
provides a really simple Makefile with no autoconf etc mess - and
while the version I will be working on and probably developing
further is JUST for using with clang and Windows-on-ARM I can
imagine having variants with say "Makefile.woa", "Makefile.rpi",
"Makefile.fedora42" and so on each distilling what the relevant OS
requires (manually). And if I get that going it could also form
a tidier basis for seeing how to use cmake that the existing
historical automake stuff...

As usual, and WoA experts who are willing to join in the debugging etc
woudl be very welcome indeed and please will they email me. Arthur


git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@7225 2bfe0521-f11c-4a00-b80e-6202646ff360